+++ src/main/java/dev/tonimatas/commands/DailyCommand.java
@@ -2,7 +2,7 @@ package dev.tonimatas.commands;
 
 import dev.tonimatas.cjda.slash.SlashCommand;
 import dev.tonimatas.config.BotFiles;
-import dev.tonimatas.api.data.DailyInfo;
+import dev.tonimatas.api.bank.DailyInfo;
 import dev.tonimatas.util.CommandUtils;
 import dev.tonimatas.util.Messages;
 import net.dv8tion.jda.api.JDA;
