+++ src/main/java/dev/tonimatas/config/data/UserData.java
@@ -1,6 +1,6 @@
 package dev.tonimatas.config.data;
 
-import dev.tonimatas.api.data.UserInfo;
+import dev.tonimatas.api.user.UserInfo;
 import dev.tonimatas.config.JsonFile;
 
 import java.util.HashMap;
